Split transaction generator into two files
